+++ src/main/java/com/demcha/compose/engine/components/style/Padding.java
@@ -1,21 +1,11 @@
 package com.demcha.compose.engine.components.style;
 
 
-import com.demcha.compose.engine.components.content.shape.Stroke;
 import com.demcha.compose.engine.components.core.Component;
-import com.demcha.compose.engine.components.core.Entity;
-import com.demcha.compose.engine.components.geometry.InnerBoxSize;
-import com.demcha.compose.engine.components.layout.RenderCoordinate;
-import com.demcha.compose.engine.components.layout.coordinator.Placement;
-import com.demcha.compose.engine.components.layout.coordinator.RenderCoordinateContext;
-import com.demcha.compose.engine.render.RenderingSystemECS;
 import lombok.extern.slf4j.Slf4j;
 
-import java.awt.*;
-import java.util.Optional;
-
 @Slf4j
-public record Padding(double top, double right, double bottom, double left) implements Component, RenderCoordinate {
+public record Padding(double top, double right, double bottom, double left) implements Component {
     public static Padding zero() {
         log.debug("Getting zero padding");
         return new Padding(0.0, 0.0, 0.0, 0.0);
@@ -48,32 +38,5 @@ public record Padding(double top, double right, double bottom, double left) impl
         result = 31 * result + Double.hashCode(bottom);
         result = 31 * result + Double.hashCode(left);
         return result;
-    }
-
-    @Override
-    public <S extends AutoCloseable>Optional<RenderCoordinateContext> renderCoordinate(Entity entity, RenderingSystemECS<S> renderingSystem) {
-        if (this.equals(zero())) {
-            log.info("Padding is zero, return empty");
-            return Optional.empty();
-        }
-        var inner = InnerBoxSize.from(entity).orElseThrow();
-        var placement = entity.getComponent(Placement.class).orElseThrow();
-        double x;
-        double y;
-        double width;
-        double height;
-        int startPage;
-        int endPage;
-
-        startPage = placement.startPage();
-        endPage = placement.endPage();
-
-        x = placement.x() + left();
-        y = placement.y()+ bottom();
-        width = inner.width();
-        height = inner.height();
-        Color color = renderingSystem.guidLineSettings().PADDING_COLOR();
-        Stroke stroke = renderingSystem.guidLineSettings().PADDING_STROKE();
-        return Optional.of(new RenderCoordinateContext(x, y, width, height, startPage, endPage,stroke, color));
     }
 }