+++ src/main/java/com/demcha/compose/engine/core/SystemECS.java
@@ -1,8 +1,0 @@
-package com.demcha.compose.engine.core;
-
-import com.demcha.compose.engine.core.EntityManager;
-
-public interface SystemECS {
-      void process(EntityManager entityManager);
-
-}