+++ src/test/java/com/demcha/mock/data/SizeData.java
@@ -1,4 +1,0 @@
-package com.demcha.mock.data;
-
-public record SizeData(double width, double height) {
-}